[skip ci] refactor: Make extensions use the dailymotion-extractor
