+++ multisrc/overrides/animestream/animexin/additional.gradle
@@ -1,4 +1,5 @@
 dependencies {
+    implementation(project(':lib-dailymotion-extractor'))
     implementation(project(':lib-okru-extractor'))
     implementation(project(':lib-gdriveplayer-extractor'))
     implementation "dev.datlag.jsunpacker:jsunpacker:1.0.1"
