+++ multisrc/overrides/animestream/luciferdonghua/additional.gradle
@@ -1,4 +1,4 @@
 dependencies {
+    implementation(project(':lib-dailymotion-extractor'))
     implementation(project(':lib-okru-extractor'))
-    implementation(project(':lib-playlist-utils'))
 }
