+++ multisrc/overrides/dooplay/donghuax/additional.gradle
@@ -1,4 +1,3 @@
 dependencies {
     implementation(project(':lib-dailymotion-extractor'))
-    implementation(project(':lib-playlist-utils'))
 }
