+++ src/ar/witanime/build.gradle
@@ -12,6 +12,7 @@ ext {
 }
 
 dependencies {
+    implementation(project(':lib-dailymotion-extractor'))
     implementation(project(':lib-dood-extractor'))
     implementation(project(':lib-okru-extractor'))
     implementation(project(':lib-mp4upload-extractor'))
