+++ src/ar/witanime/src/eu/kanade/tachiyomi/animeextension/ar/witanime/WitAnime.kt
@@ -3,7 +3,6 @@ package eu.kanade.tachiyomi.animeextension.ar.witanime
 import android.app.Application
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
-import eu.kanade.tachiyomi.animeextension.ar.witanime.extractors.DailymotionExtractor
 import eu.kanade.tachiyomi.animeextension.ar.witanime.extractors.SharedExtractor
 import eu.kanade.tachiyomi.animeextension.ar.witanime.extractors.SoraPlayExtractor
 import eu.kanade.tachiyomi.animesource.ConfigurableAnimeSource
@@ -12,6 +11,7 @@ import eu.kanade.tachiyomi.animesource.model.SAnime
 import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.ParsedAnimeHttpSource
+import eu.kanade.tachiyomi.lib.dailymotionextractor.DailymotionExtractor
 import eu.kanade.tachiyomi.lib.doodextractor.DoodExtractor
 import eu.kanade.tachiyomi.lib.mp4uploadextractor.Mp4uploadExtractor
 import eu.kanade.tachiyomi.lib.okruextractor.OkruExtractor
@@ -135,7 +135,7 @@ class WitAnime : ConfigurableAnimeSource, ParsedAnimeHttpSource() {
     private val soraPlayExtractor by lazy { SoraPlayExtractor(client) }
     private val doodExtractor by lazy { DoodExtractor(client) }
     private val sharedExtractor by lazy { SharedExtractor(client) }
-    private val dailymotionExtractor by lazy { DailymotionExtractor(client) }
+    private val dailymotionExtractor by lazy { DailymotionExtractor(client, headers) }
     private val okruExtractor by lazy { OkruExtractor(client) }
     private val mp4uploadExtractor by lazy { Mp4uploadExtractor(client) }
     private val vidBomExtractor by lazy { VidBomExtractor(client) }
@@ -159,7 +159,7 @@ class WitAnime : ConfigurableAnimeSource, ParsedAnimeHttpSource() {
             }
 
             url.contains("dailymotion") -> {
-                dailymotionExtractor.videosFromUrl(url, headers)
+                dailymotionExtractor.videosFromUrl(url)
             }
             url.contains("ok.ru") -> {
                 okruExtractor.videosFromUrl(url)
