+++ src/en/myanime/src/eu/kanade/tachiyomi/animeextension/en/myanime/Myanime.kt
@@ -4,7 +4,6 @@ import android.app.Application
 import android.content.SharedPreferences
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
-import eu.kanade.tachiyomi.animeextension.en.myanime.extractors.DailymotionExtractor
 import eu.kanade.tachiyomi.animeextension.en.myanime.extractors.YouTubeExtractor
 import eu.kanade.tachiyomi.animesource.ConfigurableAnimeSource
 import eu.kanade.tachiyomi.animesource.model.AnimeFilter
@@ -13,6 +12,7 @@ import eu.kanade.tachiyomi.animesource.model.SAnime
 import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.animesource.model.Video
 import eu.kanade.tachiyomi.animesource.online.ParsedAnimeHttpSource
+import eu.kanade.tachiyomi.lib.dailymotionextractor.DailymotionExtractor
 import eu.kanade.tachiyomi.lib.gdriveplayerextractor.GdrivePlayerExtractor
 import eu.kanade.tachiyomi.lib.okruextractor.OkruExtractor
 import eu.kanade.tachiyomi.network.GET
@@ -207,7 +207,7 @@ class Myanime : ConfigurableAnimeSource, ParsedAnimeHttpSource() {
 
                     when {
                         url.contains("dailymotion") -> {
-                            DailymotionExtractor(client).videosFromUrl(url)
+                            DailymotionExtractor(client, headers).videosFromUrl(url)
                         }
                         url.contains("ok.ru") -> {
                             OkruExtractor(client).videosFromUrl(url)
